ktr/actions: rework code - remove arg3, arg4, arg5s again. not needed anymore - ktr: add function that constructs Action() constructor arguments - simplify code massively

+++ kentauros/actions/__init__.py
@@ -33,7 +33,8 @@ class Action:
         kentauros.actions.Action.execute:
         execute Action
         """
-        pass
+        # pylint: disable=no-self-use
+        return True
 
 
 class BuildAction(Action):
@@ -41,18 +42,9 @@ class BuildAction(Action):
     kentauros.actions.BuildAction:
     action for building source package locally
     """
-    def __init__(self, package, force, arg3=None, arg4=None, arg5=None):
+    def __init__(self, package, force):
         super().__init__(package, force)
         self.type = ActionType.BUILD
-
-        if arg3 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to BuildAction. " + str(arg3), 2)
-
-        if arg4 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to BuildAction. " + str(arg4), 2)
-
-        if arg5 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to BuildAction. " + str(arg5), 2)
 
     def execute(self):
         success = self.package.builder.build()
@@ -65,18 +57,9 @@ class ChainAction(Action):
     action for getting (if neccessary), updating,
       constructing, building and uploading source package
     """
-    def __init__(self, package, force, arg3=None, arg4=None, arg5=None):
+    def __init__(self, package, force):
         super().__init__(package, force)
         self.type = ActionType.CHAIN
-
-        if arg3 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to ChainAction. " + str(arg3), 2)
-
-        if arg4 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to ChainAction. " + str(arg4), 2)
-
-        if arg5 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to ChainAction. " + str(arg5), 2)
 
     def execute(self):
         veryfied = VerifyAction(self.package, self.force).execute()
@@ -108,21 +91,13 @@ class CleanAction(Action):
     kentauros.actions.Clean:
     action for cleaning sources
     """
-    def __init__(self, package, force, arg3=None, arg4=None, arg5=None):
+    def __init__(self, package, force):
         super().__init__(package, force)
         self.type = ActionType.CLEAN
 
-        if arg3 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to CleanAction. " + str(arg3), 2)
-
-        if arg4 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to CleanAction. " + str(arg4), 2)
-
-        if arg5 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to CleanAction. " + str(arg5), 2)
-
     def execute(self):
         self.package.source.clean()
+        return True
 
 
 class ConfigAction(Action):
@@ -158,18 +133,9 @@ class ConstructAction(Action):
     kentauros.actions.Construct:
     action for building source package
     """
-    def __init__(self, package, force, arg3=None, arg4=None, arg5=None):
+    def __init__(self, package, force):
         super().__init__(package, force)
         self.type = ActionType.CONSTRUCT
-
-        if arg3 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to ConstructAction. " + str(arg3), 2)
-
-        if arg4 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to ConstructAction. " + str(arg4), 2)
-
-        if arg5 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to ConstructAction. " + str(arg5), 2)
 
     def execute(self):
         self.package.constructor.init()
@@ -190,17 +156,8 @@ class CreateAction(Action):
     kentauros.actions.CreateAction:
     action for initialising an empty package from templates
     """
-    def __init__(self, name, force=False, arg3=None, arg4=None, arg5=None):
+    def __init__(self, name, force):
         assert isinstance(name, str)
-
-        if arg3 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to CreateAction. " + str(arg3), 2)
-
-        if arg4 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to CreateAction. " + str(arg4), 2)
-
-        if arg5 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to CreateAction. " + str(arg5), 2)
 
         # copy templates from KTR_SYSTEM_DATADIR to confdir and specdir
         conf_template_orig = os.path.join(KTR_SYSTEM_DATADIR, "package.conf")
@@ -209,22 +166,34 @@ class CreateAction(Action):
         conf_template_dest = os.path.join(KTR_CONF.confdir, name + ".conf")
         spec_template_dest = os.path.join(KTR_CONF.specdir, name + ".spec")
 
-        shutil.copy2(conf_template_orig, conf_template_dest)
-        shutil.copy2(spec_template_orig, spec_template_dest)
+        self.success = True
 
-        # set name in config template
-        conf_template = ConfigParser()
-        conf_template.read(conf_template_dest)
+        if not os.path.exists(conf_template_dest) or force:
+            shutil.copy2(conf_template_orig, conf_template_dest)
+        else:
+            log(LOGPREFIX1 + name + ".conf already exists. Specify --force to overwrite.", 2)
+            self.success = False
 
-        conf_template.set("package", "name", name)
-        conf_template.write(conf_template_dest)
+        if not os.path.exists(spec_template_dest) or force:
+            shutil.copy2(spec_template_orig, spec_template_dest)
+        else:
+            log(LOGPREFIX1 + name + ".spec already exists. Specify --force to overwrite.", 2)
+            self.success = False
 
-        # initialise package
-        package = Package(name)
-        super().__init__(package, force)
+        if self.success:
+            # set name in config template
+            conf_template = ConfigParser()
+            conf_template.read(conf_template_dest)
+
+            conf_template.set("package", "name", name)
+            conf_template.write(conf_template_dest)
+
+            # initialise package
+            package = Package(name)
+            super().__init__(package, force)
 
     def execute(self):
-        pass
+        return self.success
 
 
 class ExportAction(Action):
@@ -232,21 +201,13 @@ class ExportAction(Action):
     kentauros.actions.ExportAction:
     action for exporting sources from repository
     """
-    def __init__(self, package, force, arg3=None, arg4=None, arg5=None):
+    def __init__(self, package, force):
         super().__init__(package, force)
         self.type = ActionType.EXPORT
 
-        if arg3 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to ExportAction. " + str(arg3), 2)
-
-        if arg4 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to ExportAction. " + str(arg4), 2)
-
-        if arg5 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to ExportAction. " + str(arg5), 2)
-
     def execute(self):
         self.package.source.export()
+        return True
 
 
 class GetAction(Action):
@@ -254,18 +215,9 @@ class GetAction(Action):
     kentauros.actions.GetAction:
     action for getting sources
     """
-    def __init__(self, package, force, arg3=None, arg4=None, arg5=None):
+    def __init__(self, package, force):
         super().__init__(package, force)
         self.type = ActionType.GET
-
-        if arg3 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to GetAction. " + str(arg3), 2)
-
-        if arg4 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to GetAction. " + str(arg4), 2)
-
-        if arg5 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to GetAction. " + str(arg5), 2)
 
     def execute(self):
         return self.package.source.get()
@@ -276,21 +228,15 @@ class RefreshAction(Action):
     kentauros.actions.RefreshAction:
     action for refreshing sources (clean + get)
     """
-    def __init__(self, package, force, arg3=None, arg4=None, arg5=None):
+    def __init__(self, package, force):
         super().__init__(package, force)
         self.type = ActionType.REFRESH
 
-        if arg3 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to RefreshAction. " + str(arg3), 2)
-
-        if arg4 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to RefreshAction. " + str(arg4), 2)
-
-        if arg5 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to RefreshAction. " + str(arg5), 2)
-
     def execute(self):
         self.package.source.refresh()
+        # TODO
+        # Source.refresh() needs return value from .get()
+        return True
 
 
 class StatusAction(Action):
@@ -298,22 +244,13 @@ class StatusAction(Action):
     kentauros.actions.StatusAction:
     action for displaying configuration values and available packages
     """
-    def __init__(self, package, force, arg3=None, arg4=None, arg5=None):
+    def __init__(self, package, force):
         super().__init__(package, force)
         self.type = ActionType.STATUS
 
-        if arg3 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to StatusAction. " + str(arg3), 2)
-
-        if arg4 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to StatusAction. " + str(arg4), 2)
-
-        if arg5 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to StatusAction. " + str(arg5), 2)
-
     def execute(self):
         # TODO
-        pass
+        return True
 
 
 class UpdateAction(Action):
@@ -321,18 +258,9 @@ class UpdateAction(Action):
     kentauros.actions.UpdateAction:
     action for updating sources
     """
-    def __init__(self, package, force, arg3=None, arg4=None, arg5=None):
+    def __init__(self, package, force):
         super().__init__(package, force)
         self.type = ActionType.UPDATE
-
-        if arg3 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to UpdateAction. " + str(arg3), 2)
-
-        if arg4 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to UpdateAction. " + str(arg4), 2)
-
-        if arg5 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to UpdateAction. " + str(arg5), 2)
 
     def execute(self):
         update = self.package.source.update()
@@ -344,21 +272,13 @@ class UploadAction(Action):
     kentauros.actions.UploadAction:
     action for uploading source package
     """
-    def __init__(self, package, force, arg3=None, arg4=None, arg5=None):
+    def __init__(self, package, force):
         super().__init__(package, force)
         self.type = ActionType.UPLOAD
 
-        if arg3 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to UploadAction. " + str(arg3), 2)
-
-        if arg4 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to UploadAction. " + str(arg4), 2)
-
-        if arg5 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to UploadAction. " + str(arg5), 2)
-
     def execute(self):
         self.package.uploader.upload()
+        return True
 
 
 class VerifyAction(Action):
@@ -366,18 +286,9 @@ class VerifyAction(Action):
     kentauros.actions.VerifyAction:
     action for verifying that everything is in place
     """
-    def __init__(self, package, force, arg3=None, arg4=None, arg5=None):
+    def __init__(self, package, force):
         super().__init__(package, force)
         self.type = ActionType.VERIFY
-
-        if arg3 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to VerifyAction. " + str(arg3), 2)
-
-        if arg4 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to VerifyAction. " + str(arg4), 2)
-
-        if arg5 is not None:
-            log(LOGPREFIX1 + "Superfluous argument supplied to VerifyAction. " + str(arg5), 2)
 
     def execute(self):
         # TODO
